propagation: three lax.cond branches pinned a dtype instead of following their sibling
    cylinder.py  parallel_side_branch   -> jnp.array(LARGE, dtype=jnp.float32)
    sphere.py    no_intersection_branch -> jnp.array(LARGE, dtype=jnp.float32)
    box.py       assign_off_surface     -> jnp.full((4, 3), -1, dtype=jnp.int32)

`lax.cond` requires both branches to return identical types. With jax_enable_x64 off everything
agrees BY ACCIDENT; switch x64 on and each sibling widens while these stay pinned:

    TypeError: true_fun output and false_fun output must have identical types,
               got ... float32[] vs. float64[]

The first two now take ray_origin.dtype. The third is not fixed at the pin but at its cause:
closest_face = jnp.argmin(...) is int64 under x64 and is placed into `indices` beside x_idx and
z_idx, which are already cast to int32, so the whole array promoted and the pinned branch no longer
matched. Casting closest_face makes box follow the convention cylinder and sphere already use --
every index component explicitly int32 -- which is exactly why those two survive at int32 while box
did not. All bit-identical at float32, the precision everything actually runs in.

An earlier version of this commit fixed only the first two and claimed "box.py has none". That was
wrong: the grep behind it searched for dtype=jnp.float32 and never saw an int32 pin of the same
class. Box geometries could still not be built in double precision, so "the library cannot be run
in double precision at all" was only half repaired.

The consequence of the defect is worth fixing on its own terms: float64 is how you check your own
arithmetic, and a differentiable simulator that cannot be asked whether float32 is corrupting a
gradient has to answer that question by argument instead of measurement.

tests/test_float64_runnable.py exercises all three primitives under x64 in a SUBPROCESS, since the
flag is process-global and must be set before jax touches an array, so toggling it in a live
session would contaminate every other test in the run. Each test drives BOTH branches of its cond
in the same precision, so a future pin on either side is caught.

Every fix here was checked by reverting it. All three present: 4 passed. Cylinder pin restored: 1
red. Sphere pin restored: 1 red. closest_face cast reverted: the box test goes red with the
identical-types TypeError, and restored it builds (8932, 4, 3) int32. That last check earned its
keep -- the box test was first written against a guessed signature, so it failed whether or not the
fix was present and guarded nothing. A test that has not been shown to go red is not a guard.

Replayed onto main from 3260d44, extended to box.py.

+++ lucid/propagation/box.py
@@ -371,7 +371,12 @@ def assign_sensors_to_box_grid(sensors, sensor_radius, length, width, height,
             jnp.abs(z + height/2)   # Bottom face
         ])
         
-        closest_face = jnp.argmin(dist_to_faces)
+        # .astype(int32) for the same reason cylinder.py and sphere.py cast their index
+        # components: this value is placed into `indices` beside x_idx/z_idx, which ARE cast, and
+        # `jnp.argmin` returns int64 once jax_enable_x64 is on. That promoted `indices` to int64
+        # while assign_off_surface below stays pinned int32, and `lax.cond` rejected the pair --
+        # so box geometries could not be built in double precision at all.
+        closest_face = jnp.argmin(dist_to_faces).astype(jnp.int32)
         min_distance = jnp.min(dist_to_faces)
         
         # Check if sensor is close enough to any face
